Fix reflection of inherited Repository methods that start with find[One]By
When using inheritance within Repositories, and using custom findBy methods that
don't follow the naming scheme of model-properties, the previous code broke.

Suppose a structure like this:
  PostRepository

  SpecialPostRepository extends PostRepository {
    public function findByNonExistingProperty(){}
  }

Phpstan checks for the existence of every method in a child class for all parent classes
to ensure their arguments are compatible.
This leads to a hasMethod call for 'findByNonExistingProperty' on PostRepository.
Note that this class does not have a native implementation, so previously the FindByMethodReflection
was returned, despite the fact, that this property doesn't even exist.

Now we properly check for its existence beforehand and return false for hasMethod in that case