Do not fail on calls to MyRepository->findBySomePropNotInModel()
Calling a findByXXX method on a repository tries to get the type of
the property from the model. But the reposiory does also allow findBy
methods for values that are not defined in the model but only in the
database. So if the model does not have the certain property we tread
it as type mixed.