Add possibility to define array return type for Repository->findAll
The RepositoryFindAllDynamicReturnTypeExtension does now check where
the findAll method is defined. If it is not overwritten but defined
in the native Repository class, the default behavior, returning a
QueryResultInterface, will be used. If the findAll method is overwritten,
the declaration from that overwrite will be used.
The Repository.stub has been adjusted to now also support and array as
return type for findAll.