Allow more complex phpdoc types in mappings and add more tests
The types defined in the `parameters.typo3.contextApiGetAspectMapping`
and `parameters.typo3.requestGetAttributeMapping` can now be more complex
phpDoc types. This is useful if e.g. a request argument can be nullable
or of diffent type.

Tests have been added to test the behavior.

Beside that a validation rule has been added that gives a hint
if a request attribute mapping for phpstan is missing.