[BUGFIX] Fix fatal error in QueryInterfaceDynamicReturnTypeExtension
In some rare cases (only in combination with rector?) the
QueryInterfaceDynamicReturnTypeExtension failed with an exception
as it has been tried to convert a Repository name to a model name
even if the QueryInterface was not used in a Repository.
It is checked if we're in the scope of a repository and if not
we use a mixed type as we could not find out which type exactly
it is.