Refactor for PHPStan 2 Compatibility
Refactor the entire package to ensure compatibility with PHPStan 2.
To reduce effort, the package is now only compatible with TYPO3 v13,
and compatibility with TYPO3 v12 and v11 has been removed.
Older TYPO3 versions will continue to be supported with version 1 of
the package. If support for TYPO3 v12 in version 2 is required,
it can still be implemented if necessary.