Fix invalid interface implementation by dynamic return type
The object storage offsetGet() is not compatible with the
ArrayAccess interface but does some TYPO3 sepcific stuff.
We can't set the correct return type via a stub as phpstan
will complain that the return type is not compatible with
the interface.
To work around this we have to implement a dynamic return
type extension again.